% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{cites_codes}
\alias{cites_codes}
\title{CITES Code Values}
\usage{
cites_codes()
}
\value{
A tibble with fields and descriptions
}
\description{
This function returns a data frame with descriptions of all the code values
used in [cites_data()].  This is useful for lookup
as well as merging with the data for more descriptive summaries.
}
\details{
These values are drawn from
["A guide to using the CITES Trade Database"](https://github.com/ecohealthalliance/cites/tree/master/inst/extdata),
from the CITES website.

\if{html}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
    if(citesdb:::is_js_ok()) {
      mytext <- cites:::rd_datatable(cites::cites_codes())
    } else {
      mytext <- c('In RStudio help, this help file includes a searchable table of values if you install the DT package')
    }
    mytext
  }
}

\if{text,latex}{The HTML version of this help file includes a searchable table of the CITES codes.}
}
\seealso{
[cites_metadata()] [cites_data()]
}
