% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{cites_db}
\alias{cites_db}
\title{The local CITES database}
\usage{
cites_db(dbdir = cites_path())
}
\arguments{
\item{dbdir}{The location of the database on disk.  Defaults to
`citesdb` under [MonetDBLite::MonetDBLite], or `getOption("CITES_DB_DIR")`.}
}
\value{
A MonetDBLite DBI connection
}
\description{
Returns a connecction to the local CITES database.  This is a DBI-complient
[MonetDBLite::MonetDBLite]() database.
}
\examples{
cites_db()
}
