% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onattach.R
\name{cites_db_delete}
\alias{cites_db_delete}
\title{Remove the local CITES database}
\usage{
cites_db_delete()
}
\description{
Deletes all tables from the local database
}
\examples{
\donttest{
\dontrun{
cites_db_delete()
}
}
}
