% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{cites_db_download}
\alias{cites_db_download}
\title{Title}
\usage{
cites_db_download(codes = NULL, db = cites_db(), new_only = TRUE,
  verbose = TRUE, temp_dir = tempdir(), cleanup = TRUEs)
}
\arguments{
\item{cleanup}{}
}
\description{
Title
}
