% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{cites_disconnect}
\alias{cites_disconnect}
\title{Disconnect from the CITES database}
\usage{
cites_disconnect()
}
\description{
A utility function for disconnecting from the database.
}
\examples{
cites_disconnect()
}
