% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{cites_metadata}
\alias{cites_metadata}
\alias{metadata}
\title{CITES Field Descriptions}
\usage{
cites_metadata()
}
\value{
A tibble with field, code, and code description
}
\description{
This function returns a data frame field descriptions for [cites_data()].
}
\details{
This information is taken from
["A guide to using the CITES Trade Database"](https://github.com/ecohealthalliance/cites/tree/master/inst/extdata),
from the CITES websites.

\if{html}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
    if (citesdb:::is_js_ok()) {
      mytext <- cites:::rd_datatable(cites::cites_metadata())
    } else {
      mytext <- cites:::tabular(cites::cites_metadata())
    }
    mytext
  }
}

\if{text,latex}{ \Sexpr[echo=FALSE, results=rd, stage=build]{cites:::tabular(cites::cites_metadata())}}
}
\seealso{
[cites_codes()] [cites_data()] [cites_parties()]
}
