% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pane.R
\name{cites_pane}
\alias{cites_pane}
\title{Open CITES database connecction pane in RStudio}
\usage{
cites_pane()
}
\description{
This function launches the RStudio "Connection" pane to interactively
explore the database.
}
\examples{
if (!is.null(getOption("connectionObserver")))
  cites_pane()
}
