% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{cites_parties}
\alias{cites_parties}
\alias{parties}
\title{Parties to the CITES treaty.}
\usage{
cites_parties()
}
\value{
A tibble
}
\description{
This function returns a data frame witha list of countries party to CITES
and their date of joining the treaty.
}
\details{
\if{html}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
    if (citesdb:::is_js_ok()) {
      mytext <- cites:::rd_datatable(cites::cites_parties())
    } else {
      mytext <- cites:::tabular(cites::cites_parties())
    }
    mytext
  }
}

\if{text,latex}{ \Sexpr[echo=FALSE, results=rd, stage=build]{cites:::tabular(cites::cites_metadata())}}
}
\seealso{
[cites_codes()] [cites_metadata()]
}
