% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onattach.R
\name{cites_status}
\alias{cites_status}
\title{Get the status of the current local CITES database}
\usage{
cites_status(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to print a status message}
}
\value{
TRUE if the database exists, FALSE if it is not detected. (invisible)
}
\description{
Get the status of the current local CITES database
}
\examples{
cites_status()
}
