% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citesdb-package.R
\docType{package}
\name{citesdb-package}
\alias{citesdb}
\alias{citesdb-package}
\title{The CITES database package}
\description{
**citesdb** provides over 40 years of trade data from the [Convention on
International Trade in Endangered Species of Wild Fauna and Flora](https://www.cites.org).
These data on imports and exports of CITES annex species are provided by
CITES membere countries to the CITES secratariat.
}
\details{
This package allows you to download the full database of histric transactions
to anayze locally on your computer. Because of the large size of this data,
the **citesdb** package stores it as an on-disk [MonetDB](https://www.monetdb.org/)
that can be queried without loading into RAM.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecohealthalliancec/citesdb}
  \item \url{https://www.cites.org/}
  \item Report bugs at \url{https://github.com/ecohealthalliancce/citesdb/issues}
}

}
\author{
Noam Ross, Evan Eskew, and Carlos Zambrana-Torrelio, [EcoHealth Alliance](https://www.ecohealthalliance.org/)
}
\keyword{internal}
