% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{cites_db}
\alias{cites_db}
\title{The local CITES database}
\usage{
cites_db(dbdir = cites_path())
}
\arguments{
\item{dbdir}{The location of the database on disk. Defaults to
\code{citesdb} under \code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}}, or the environment variable \code{CITES_DB_DIR}.}
}
\value{
A MonetDBLite DBI connection
}
\description{
Returns a connection to the local CITES database. This is a DBI-compliant
\code{\link[MonetDBLite:MonetDBLite]{MonetDBLite::MonetDBLite()}} database connection. When using \strong{dplyr}-based
workflows, one typically accesses tables with functions such as
\code{\link[=cites_shipments]{cites_shipments()}}, but this function lets one interact with the database
directly via SQL.
}
\examples{
if (cites_status()) {
  library(DBI)
  dbListTables(cites_db())
  parties <- dbReadTable(cites_db(), "cites_parties")
  dbGetQuery(
   cites_db(),
   'SELECT "Taxon", "Importer" FROM cites_shipments WHERE "Year" = 1976 LIMIT 100;'
   )
}
}
