% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{cites_db_download}
\alias{cites_db_download}
\title{Download the CITES database to your local computer}
\usage{
cites_db_download(tag = NULL, destdir = tempdir(), cleanup = TRUE,
  verbose = interactive())
}
\arguments{
\item{tag}{What release tag of data to download. Defaults to the most recent}

\item{destdir}{Where to download the compressed file.}

\item{cleanup}{Whether to delete the compressed file after loading into the database.}

\item{verbose}{Whether to display messages and download progress}
}
\description{
This command downloads the CITES shipments database and populates a local
database.  The download is large (>80MB), and the database will be at least
twice that on disk.  During import over 1GB of disk space may be used.
}
\examples{
cites_db_download()
}
