% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{cites_metadata}
\alias{cites_metadata}
\alias{metadata}
\alias{cites_codes}
\alias{cites_parties}
\title{Cites shipment metadata}
\usage{
cites_metadata()

cites_codes()

cites_parties()
}
\value{
A tibble of metadata
}
\description{
The CITES database also includes tables of column-levelel metadata and
meanings of codes in colums, as well as a listing of CITES parties.
convenience functions acccess these tables. As they are small, the functions
collect the into R session memory, rather than returning a remote table.

This information is drawn from
\href{https://trade.cites.org/cites_trade_guidelines/en-CITES_Trade_Database_Guide.pdf}{"A guide to using the CITES Trade Database"},
from the CITES website. More information the the shipment-level data can be
found in the \link{guidance} help file.
}
\examples{
cites_metadata()
cites_codes()
cites_shipments()

# For remote connections to these tables,
# access the database directly:
dplyr::tbl(cites_db(), "cites_metadata")
dplyr::tbl(cites_db(), "cites_codes")
dplyr::tbl(cites_db(), "cites_parties")
}
