% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{cites_metadata}
\alias{cites_metadata}
\alias{metadata}
\alias{cites_codes}
\alias{cites_parties}
\title{CITES shipment metadata}
\usage{
cites_metadata()

cites_codes()

cites_parties()
}
\value{
A tibble of metadata
}
\description{
The CITES database also includes tables of column-level metadata and
meanings of codes in columns, as well as a listing of CITES Parties/country abbreviations.
Convenience functions access these tables. As they are small, the functions
collect the tables into R session memory, rather than returning a remote table.

This information is drawn from
\href{https://trade.cites.org/cites_trade_guidelines/en-CITES_Trade_Database_Guide.pdf}{"A guide to using the CITES Trade Database"},
from the CITES website. More information on the shipment-level data can be
found in the \link{guidance} help file.
}
\examples{
if (cites_status()) {
  library(dplyr)

  # Look at the field definitions for cites_shipments()
  cites_metadata()
  # View the codes used for purpose of shipment
  cites_codes() \%>\%
   filter(field == "Purpose")

  # See the most recent countries to join CITES
  cites_parties() \%>\%
    arrange(desc(date)) \%>\%
    head(10)

  # See countries or locations with non-standaard or outdated ISO codes
  cites_parties() \%>\%
    filter(former_code | non_ISO_code)

  # For remote connections to these tables,
  # access the database directly:
  dplyr::tbl(cites_db(), "cites_metadata")
  dplyr::tbl(cites_db(), "cites_codes")
  dplyr::tbl(cites_db(), "cites_parties")
}
}
