% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citesdb-package.R
\docType{package}
\name{citesdb-package}
\alias{citesdb}
\alias{citesdb-package}
\title{The CITES database package}
\description{
\strong{citesdb} provides over 40 years of trade data from the \href{https://www.cites.org}{Convention on International Trade in Endangered Species of Wild Fauna and Flora}.
These data on imports and exports of CITES annex species are provided by
CITES member countries to the CITES secretariat.
}
\details{
This package allows you to download the full database of historic transactions
to analyze locally on your computer. Because of the large size of this data,
the \strong{citesdb} package stores it as an on-disk \href{https://www.monetdb.org/}{MonetDB}
that can be queried without loading into RAM.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecohealthalliancec/citesdb}
  \item \url{https://www.cites.org/}
  \item Report bugs at \url{https://github.com/ecohealthalliancce/citesdb/issues}
}

}
\author{
Noam Ross, Evan Eskew, and Carlos Zambrana-Torrelio, \href{https://www.ecohealthalliance.org/}{EcoHealth Alliance}
}
\keyword{internal}
