% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citesdb-package.R
\docType{package}
\name{citesdb-package}
\alias{citesdb}
\alias{citesdb-package}
\title{citesdb: Historical CITES trade data served via databae}
\description{
It does some stuff
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecohealthalliancec/citesdb}
  \item Report bugs at \url{https://github.com/ecohealthalliancce/citesdb/issues}
}

}
\author{
\strong{Maintainer}: Noam Ross \email{ross@ecohealthalliance.org}

Authors:
\itemize{
  \item Evan Eskew \email{eskew@ecohealthalliance.org}
  \item Carlos Zambrana-Torrelio \email{zambrana@ecohealthalliance.org}
}

Other contributors:
\itemize{
  \item EcoHealth Alliance [copyright holder]
}

}
\keyword{internal}
